Changed to 53-Bit-Id (Javascript MAX SAFE INTEGER)

+++ Pillepalle1.StatefulApi/src/Customers.Application/Cqrs/Commands/DeleteCustomerCmd.cs
@@ -2,7 +2,7 @@ namespace Customers.Application.Cqrs.Commands;
 
 public class DeleteCustomerCmd : ARequestBase<OneOf<int,Problem>>
 {
-    public required int Id { init; get; }
+    public required long Id { init; get; }
 }
 
 public class DeleteCustomerCmdValidator : AbstractValidator<DeleteCustomerCmd>
